mlonmcu/setup/tasks/cmake.py: cleanup src dir as well

+++ mlonmcu/setup/tasks/cmake.py
@@ -148,17 +148,26 @@ def _validate_cmake_clean(context: MlonMcuContext, params={}):
         return False
     user_vars = context.environment.vars
     keep_build_dir = user_vars.get("cmake.keep_build_dir", False)
-    return not keep_build_dir
+    keep_src_dir = user_vars.get("cmake.keep_src_dir", False)
+    return (not keep_build_dir) or (not keep_src_dir)
 
 
-@Tasks.needs(["cmake.exe", "cmake.build_dir"])
-@Tasks.removes(["cmake.build_dir"])  # TODO: implement
+@Tasks.needs(["cmake.exe", "cmake.src_dir", "cmake.build_dir"])
+@Tasks.removes(["cmake.build_dir", "cmake.src_dir"])  # TODO: implement
 @Tasks.validate(_validate_cmake_clean)
 @Tasks.register(category=TaskType.TARGET)
 def clean_cmake(
     context: MlonMcuContext, params=None, rebuild=False, verbose=False, threads=multiprocessing.cpu_count()
 ):
     """Cleanup CMake build dir."""
-    cmakeBuildDir = context.cache["cmake.build_dir"]
-    shutil.rmtree(cmakeBuildDir)
-    del context.cache["cmake.build_dir"]
+    user_vars = context.environment.vars
+    keep_build_dir = user_vars.get("cmake.keep_build_dir", False)
+    if not keep_build_dir:
+        cmakeBuildDir = context.cache["cmake.build_dir"]
+        shutil.rmtree(cmakeBuildDir)
+        del context.cache["cmake.build_dir"]
+    keep_src_dir = user_vars.get("cmake.keep_src_dir", False)
+    if not keep_src_dir:
+        cmakeSrcDir = context.cache["cmake.src_dir"]
+        shutil.rmtree(cmakeSrcDir)
+        del context.cache["cmake.src_dir"]
